dia 14

+++ src/test/java/dev/serverest/stepdef/UsuarioStepDef.java
@@ -1,6 +1,8 @@
 package dev.serverest.stepdef;
 
+import dev.serverest.api.applicationAPI.LoginAPI;
 import dev.serverest.api.applicationAPI.UsuariosAPI;
+import dev.serverest.pojo.Login;
 import dev.serverest.pojo.Usuario;
 import dev.serverest.pojo.Usuarios;
 import io.cucumber.datatable.DataTable;
@@ -16,9 +18,9 @@ import static dev.serverest.api.applicationAPI.UsuariosAPI.generateRandomUser;
 import static dev.serverest.utils.FakerUtils.*;
 
 public class UsuarioStepDef {
-    private Usuario responseAsClass;
-    private Usuarios responseAsClassList;
-    private Usuario requestUsuario;
+    private Usuario responseAsClass = new Usuario();
+    private Usuarios responseAsClassList = new Usuarios();
+    private Usuario requestUsuario = new Usuario();
     private Response response;
 
     @Dado("^que quero criar um usuario (normal|admin)$")
@@ -28,41 +30,52 @@ public class UsuarioStepDef {
 
     @Quando("eu fizer um POST na API de cadastro")
     public void eu_fizer_um_POST_na_API_de_cadastro() {
-       response = UsuariosAPI.post(requestUsuario);
+        response = UsuariosAPI.post(requestUsuario);
         responseAsClass = response.as(Usuario.class);
     }
 
     @Entao("o status code sera {int}")
     public void o_status_code_sera(Integer status) {
-        Assert.assertEquals(response.statusCode(),status);
+        Assert.assertEquals(response.statusCode(), status);
 
-        UsuariosAPI.delete(responseAsClass.getId());
+        if (responseAsClass != null) {
+            UsuariosAPI.delete(responseAsClass.getId());
+        }
     }
 
-    @Dado("tenho uma conta ja cadastrada")
+    @Dado("que eu tenho uma conta ja cadastrada")
     public void tenho_Uma_Conta_Ja_Cadastrada() {
         requestUsuario = generateRandomUser();
         response = UsuariosAPI.post(requestUsuario);
-        
+        responseAsClass = response.as(Usuario.class);
     }
 
     @Quando("eu fizer um PUT na API de cadastro")
     public void eu_Fizer_Um_PUT_Na_API_De_Cadastro() {
-        responseAsClass = response.as(Usuario.class);
         requestUsuario = generateRandomUser();
-        response = UsuariosAPI.put(requestUsuario,responseAsClass.getId());
+        response = UsuariosAPI.put(requestUsuario, responseAsClass.getId());
     }
 
     @Quando("eu fizer um GET na API de cadastro com o id")
     public void eu_Fizer_Um_GET_Na_API_De_Cadastro() {
-        responseAsClass = response.as(Usuario.class);
-        response = UsuariosAPI.get(responseAsClass.getId());
+        response = UsuariosAPI.get("_id", responseAsClass.getId());
     }
 
     @Entao("a quantidade de usuarios achados sera {int}")
     public void a_Quantidade_De_Usuarios_Achados_Sera(int quantidade) {
         responseAsClassList = response.as(Usuarios.class);
-        Assert.assertEquals(responseAsClassList.getQuantidade(),quantidade);
+        Assert.assertEquals(responseAsClassList.getQuantidade(), quantidade);
         UsuariosAPI.delete(responseAsClass.getId());
+    }
+
+    @Dado("que eu tenho uma conta nao cadastrada")
+    public void queEuTenhoUmaContaNaoCadastrada() {
+        requestUsuario = generateRandomUser();
+    }
+
+    @Quando("eu fizer um POST na API de login")
+    public void euFizerUmPOSTNaAPIDeLogin() {
+        Login requestLogin = LoginAPI.loginBuilder(requestUsuario.getEmail(), requestUsuario.getPassword());
+        response = LoginAPI.post(requestLogin);
     }
 }
